Update index.ts
Correction of xFraction variable calculation.

I propose to calculate the distance in latitude from the edge of the box to the center and multiply by 2, because with a large box (more than half the length of the latitude), the distance is not calculated correctly (the shortest distance from edge to edge will envelope the Earth from the other side and will not be the length of the box).

I also propose to calculate the distance in latitude, located in the center of the height of the box, because Now it is calculated on the lower edge (south), which gives a very small number of points for large boxes (at a height from the equator to the pole), because at the northern and southern latitudes their length decreases with distance from the equator. Also, with the same small boxes (one in the northern hemisphere, the other in the southern), the difference in result will be visible. In this regard, I think it will be fair to consider the latitude located in the center of the box.

And you also need to revise the tests for this module.

Thank.